pyproject.toml
[tool.poetry]
name = "hexlet-calendar"
version = "0.1.0"
description = "tool for getting hexlet calendar info and shipping it somewhere (like to Telegram bot)"
authors = ["Danil Kalistov <simr0xxorz@gmail.com>"]
readme = "README.md"
packages = [{include = "calendarbot"}]

[tool.poetry.dependencies]
python = "^3.9"
google-api-python-client = "^2.74.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.8.0"
requests = "^2.28.2"
typing-extensions = "^4.4.0"
python-dotenv = "^0.21.1"
sqlalchemy = "^2.0.2"
psycopg2-binary = "^2.9.5"

[tool.poetry.scripts]
start = "calendarbot.start:main"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
